FIX(client): Save shortcuts in a single DB transaction
The server-specific shortcuts, which are saved to the DB, were
previously saved in two steps: First deleting all old shortcuts that are
stored in the DB for that server and then inserting the new shortcuts
afterwards.

However, if Mumble was to crash or forcefully quit for another reason
between these two steps, the server-specific shortcuts would be lost
completely (the old ones have been intentionally deleted from the DB and
the new ones haven't been saved yet).

Therefore, this commit now wraps these two steps into a single
transaction, which will execute atomically. So either the full
transaction succeeds (in which case the new shortcuts are saved as
expected) or it fails (or is aborted) in which case it will undo all
changes made so far (thus, in this case at least the old shortcuts that
have been stored previously are retained).

+++ src/mumble/Database.cpp
@@ -583,6 +583,15 @@ QList< Shortcut > Database::getShortcuts(const QByteArray &digest) {
 void Database::setShortcuts(const QByteArray &digest, const QList< Shortcut > &shortcuts) {
 	QSqlQuery query(db);
 
+	if (!db.transaction()) {
+		const QSqlError error(QSqlDatabase::database().lastError());
+		qWarning() << "Database: Unable to start transaction for saving shortcuts" << error.nativeErrorCode()
+				   << error.text();
+		qWarning() << "-> We'll rather not save them at all than risk potentially losing all previous shortcuts";
+
+		return;
+	}
+
 	query.prepare(QLatin1String("DELETE FROM `shortcut` WHERE `digest` = ?"));
 	query.addBindValue(digest);
 	execQueryAndLogFailure(query);
@@ -613,6 +622,8 @@ void Database::setShortcuts(const QByteArray &digest, const QList< Shortcut > &s
 			execQueryAndLogFailure(query);
 		}
 	}
+
+	db.commit();
 }
 
 const QMap< QString, QString > Database::getFriends() {
